created simplefilter class to encapsulate all filtering functionality. made bots active by default with toggle moderation mode active (moderating) or passive (filter only). rooms being moderated by bot are stored in dict as room:active_mode(bool)
